pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sr15-analytics"
version = "1.0.0"
description = "SR15 PIB × HS Code × KLU Analytics Dashboard"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.0.0",
    "dash-cytoscape>=1.0.0",
    "flask-caching>=2.3.0",
    "plotly>=6.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pages*", "utils*"]
